+++ src/CruiseProcessing.Core/Models/RegressionGroups.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace CruiseProcessing.Models
+{
+    public class RegressGroups
+    {
+        public string rgSpecies { get; set; }
+        public string rgProduct { get; set; }
+        public string rgLiveDead { get; set; }
+        public int rgSelected { get; set; }
+    }   //  end RegressGroups
+}
