+++ src/CruiseProcessing.Core/Output/OutputBLM.cs
@@ -178,14 +178,13 @@ namespace CruiseProcessing
                     numOlines = 0;
                     completeHeader = createCompleteHeader();
                     fieldLengths = new int[] { 1, 4, 9, 9, 9, 9, 9, 9, 11, 8, 9, 11, 7, 11, 7, 9 };
-                    ArrayList justSpecies = DataLayer.GetJustSpecies("Tree");
+                    var justSpecies = DataLayer.GetDistinctTreeSpeciesCodes();
                     //  load log DIBs into output list
                     List<LogStockDO> justDIBs = DataLayer.getLogDIBs();
                     //  process by species
-                    foreach (object js in justSpecies)
+                    foreach (string currSP in justSpecies)
                     {
                         LoadLogDIBclasses(justDIBs, listToOutput);
-                        currSP = Convert.ToString(js);
                         extraLine = "    SPECIES:  ";
                         extraLine += currSP.PadLeft(6, ' ');
                         SetReportTitles(currentTitle, 6, 0, 0, reportConstants.FCTO, "");
