+++ src/CruiseProcessing/ValueEquations.cs
@@ -60,22 +60,26 @@ namespace CruiseProcessing
             }   //  endif list is empty
 
             //  Fill lists at bottom with unique species and primary products
-            ArrayList justSpecies = DataLayer.GetJustSpecies("Tree");
-            for (int n = 0; n < justSpecies.Count; n++)
-                speciesList.Items.Add(justSpecies[n].ToString());
+            var justSpecies = DataLayer.GetDistinctTreeSpeciesCodes();
+            foreach(var sp in justSpecies)
+            {
+                speciesList.Items.Add(sp);
+            }
 
             //  If there are no species/products in tree default values, it's wrong
             //  tell user to check the file design in CSM --  June 2013
-            if (justSpecies.Count == 0)
+            if (!justSpecies.Any())
             {
                 MessageBox.Show("No species/product combinations found in Tree records.\nPlease enter tree records before continuing.", "ERROR", MessageBoxButtons.OK, MessageBoxIcon.Error);
                 Close();
                 return -1;
             }   //  endif
 
-            ArrayList justProducts = DataLayer.GetJustPrimaryProduct();
-            for (int n = 0; n < justProducts.Count; n++)
-                primaryProdList.Items.Add(justProducts[n].ToString());
+            var justProducts = DataLayer.GetDistincePrimaryProductCodes();
+            foreach(var prod in  justProducts)
+            {
+                primaryProdList.Items.Add(prod);
+            }
 
             regionNum.Enabled = false;
             equationNumber.Enabled = false;
